Ensured that NaN direction parameter in filterStations returns an error
In cases when user enters a non-numerical string as a direction parameter, it will end up as a NaNat line 59 of server.ts and be passed to filterStationsData() as such. Letting it enter the 'if' block at line 19 of filterData.ts will prevent running line 66-70 of filterData.ts and return an invalid parameter error as it should.